+++ backend/associados/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class AssociadosConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "associados"
+    name = "backend.associados"
     verbose_name = "Gestão de Associados"